refactor: improve code quality across proxy, websocket, and http modules
- SocketProxy.c: Extract helper functions for validation and string copying
- SocketWS-handshake.c: Add safe snprintf with overflow checking
- SocketWS-frame.c: Refactor frame parsing for improved safety
- SocketHTTPClient/Server: Minor improvements and fixes
- Tests: Update integration tests for improved coverage
- Docs: Regenerate Doxygen documentation
- CI: Add local CI script and TSAN suppression file

+++ src/fuzz/fuzz_synprotect.c
@@ -134,13 +134,13 @@ generate_cidr (char *buf, size_t bufsize, const uint8_t **data, size_t *remainin
       snprintf (buf, bufsize, "%u.%u.%u.%u/%u", read_byte (data, remaining),
                 read_byte (data, remaining), read_byte (data, remaining),
                 read_byte (data, remaining),
-                read_byte (data, remaining) % 33);
+                (unsigned)(read_byte (data, remaining) % 33));
       break;
 
     case 1: /* Valid IPv6 CIDR */
       snprintf (buf, bufsize, "%x:%x::%x/%u", read_byte (data, remaining),
                 read_byte (data, remaining), read_byte (data, remaining),
-                read_byte (data, remaining) % 129);
+                (unsigned)(read_byte (data, remaining) % 129));
       break;
 
     case 2: /* Invalid CIDR */
